Перенес взаимодействие с МосПолитехом к эндпоинтам Telegram
